Add func to add a todo into list when newTodoing
It was a tricky part because I had to find a way to have a list of
allLists with just a name. So I used the Array.find() function.
I created a special function to find a list inside the allLists,
cleverly named getListFromAllLists. But I will need to create another
one for todos inside alist, surely inside the List class.