Add expandable functionnality to todos
I think I did it the hard way using eventListeners that creates DOM
elements on click, and removing them on another click. One tricky part
was to remove the eventListener after use (the last line of each
expandable or shrinkable functions). I saw on W3School that there was
a simplier way to make collapsable divs toggling the display: none of
some part of your divs to display: block (clever way). I also didn't
think it would be easy to just use the same variables inside the
eventListeners (I mean the todo and todoEl), but I guess when triggered
it refers to the place it was established and so finds the right todo
object to create the divs from. Anyway I'm happy I found a way.