Add delete button on expandable content
I had trouble with my iteration through my lists that it was removing
objects in every lists even if it wasn't part of it. In fact it was
removing the last one if it didi'nt exist. Si I had to go back to the
definition of the removeTodo method of List objects, to remember that
indexOf returns -1 if it doesn't find what you're looking for, hence the
last object beeing removed (I use splice with the index). Kinf of proud
of myself to look for the problem, in each bit of code invoked !