Create and prepare localeSTorageLogic file
I'm starting to die with finding the logic to store in localStorage...
I'm thinking of going through every list and creating key/value pairs
with JSON.stringifying the todo objects. But for that I also need to add
a parameter to Todo class idnicating the list name it is in.