Finish retrieveStorage fn and so finish storage
Yeeeees I've finally did it !! I really was stuck on that final every()
function to check that it was a new list to add. Now it works just fine,
I don't have a DOM function to remove a list but if it is empty, it
disappear after a refresh because there's no todo stored with the list's
name so it doesn't get created again.
I'm really proud to have mastered the storage functionnality, but at the
same time I'm kind of disappointed at my code. There's a lot that could
be better, but I guess that's the steps of learning !